Reorganize tests by feature; add provider contract and modify-if-passed e2e tests
Merge the seven coverage-push/issue-numbered test files into feature-named
files so tests for any command are findable. Extract shared mock/format
fixtures into tests/integration/conftest.py. Add a provider-parity contract
suite (JsonProvider + ClickUpClient) and end-to-end tests proving
--description "" clears through the real JSON store. Tighten exit-code
assertions from !=0 to the specific contract codes.
