+++ tests/integration/test_cli_commands.py
@@ -1,12 +1,22 @@
-"""Integration tests for CLI commands."""
+"""Integration tests for CLI commands (config, help, and misc entry points).
+
+Consolidates tests from the original test_cli_commands.py, plus config tests
+formerly in test_command_coverage.py and test_more_coverage.py.
+"""
+
+from __future__ import annotations
 
 import json
 import tempfile
+from pathlib import Path
 from unittest.mock import AsyncMock, Mock, patch
 
 from typer.testing import CliRunner
 
 from clickup.cli.main import app
+from clickup.core.models import User
+
+from .conftest import make_mock_ctx
 
 runner = CliRunner()
 
@@ -213,3 +223,206 @@ def test_config_help():
     assert "Configuration" in result.stdout
     assert "set-token" in result.stdout
     assert "show" in result.stdout
+
+
+# =============================================================================
+# config: from test_command_coverage.py
+# =============================================================================
+
+
+def test_config_set_client_id():
+    with tempfile.TemporaryDirectory() as tmpdir:
+        with patch("clickup.core.config.Path.home", return_value=Path(tmpdir)):
+            result = runner.invoke(app, ["config", "set-client-id", "client_123"])
+            assert result.exit_code == 0
+            data = json.loads(result.stdout)
+            assert data["key"] == "client_id"
+
+
+def test_config_set_client_secret():
+    with tempfile.TemporaryDirectory() as tmpdir:
+        with patch("clickup.core.config.Path.home", return_value=Path(tmpdir)):
+            result = runner.invoke(app, ["config", "set-client-secret", "secret_456"])
+            assert result.exit_code == 0
+
+
+def test_config_get_unset():
+    """`config get` on missing key emits null value."""
+    result = runner.invoke(app, ["config", "get", "nonexistent_key_xyz"])
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data["value"] is None
+
+
+def test_config_show_json():
+    result = runner.invoke(app, ["config", "show"])
+    assert result.exit_code == 0
+
+
+def test_config_clean_no_unknowns():
+    """Clean is a no-op when there are no unknown keys."""
+    result = runner.invoke(app, ["config", "clean"])
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data["unknown_keys"] == 0
+
+
+def test_config_clean_dry_run():
+    """Plant an unknown key, then dry-run shouldn't remove it or refuse."""
+    from clickup.core import Config
+
+    cfg = Config()
+    cfg_path = Path(cfg._get_config_path())
+    cfg_path.parent.mkdir(parents=True, exist_ok=True)
+    cfg_path.write_text(json.dumps({"default_team_id": "T1", "_garbage": "xx"}))
+
+    result = runner.invoke(app, ["config", "clean", "--dry-run"])
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data["dry_run"] is True
+    assert data["would_remove"] == 1
+
+
+def test_config_clean_with_force_removes():
+    from clickup.core import Config
+
+    cfg = Config()
+    cfg_path = Path(cfg._get_config_path())
+    cfg_path.parent.mkdir(parents=True, exist_ok=True)
+    cfg_path.write_text(json.dumps({"default_team_id": "T1", "_garbage": "xx"}))
+
+    result = runner.invoke(app, ["config", "clean", "--force"])
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data["removed"] == 1
+
+
+def test_config_set_default_list():
+    result = runner.invoke(app, ["config", "set-default-list", "myalias", "12345"])
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data["alias"] == "myalias"
+
+
+def test_config_set_default_list_remove():
+    runner.invoke(app, ["config", "set-default-list", "myalias", "12345"])
+    result = runner.invoke(app, ["config", "set-default-list", "--remove", "myalias"])
+    assert result.exit_code == 0
+    data = json.loads(result.stdout)
+    assert data["action"] == "removed"
+
+
+def test_config_set_default_list_remove_missing_errors():
+    result = runner.invoke(app, ["config", "set-default-list", "--remove", "nonexistent"])
+    assert result.exit_code == 1
+
+
+def test_config_set_default_list_no_id_errors():
+    result = runner.invoke(app, ["config", "set-default-list", "myalias"])
+    assert result.exit_code == 1
+    assert "list_id" in result.output
+
+
+# =============================================================================
+# config validate / whoami — from test_more_coverage.py
+# =============================================================================
+
+
+@patch("clickup.cli.commands.config.get_provider")
+def test_config_validate_valid_token(mock_client_cls):
+    mock_client = AsyncMock()
+    mock_client.validate_auth.return_value = (
+        True,
+        "Authenticated",
+        User(id=1, username="evan", email="e@x.com", color="#fff", profilePicture="x"),
+    )
+    mock_client_cls.return_value = make_mock_ctx(mock_client)
+
+    with tempfile.TemporaryDirectory() as tmpdir:
+        with patch("clickup.core.config.Path.home", return_value=Path(tmpdir)):
+            runner.invoke(app, ["config", "set-token", "pk_test"])
+            result = runner.invoke(app, ["config", "validate"])
+            assert result.exit_code == 0
+            assert "evan" in result.output
+
+
+@patch("clickup.cli.commands.config.get_provider")
+def test_config_validate_invalid_token(mock_client_cls):
+    mock_client = AsyncMock()
+    mock_client.validate_auth.return_value = (False, "bad token", None)
+    mock_client_cls.return_value = make_mock_ctx(mock_client)
+
+    with tempfile.TemporaryDirectory() as tmpdir:
+        with patch("clickup.core.config.Path.home", return_value=Path(tmpdir)):
+            runner.invoke(app, ["config", "set-token", "pk_bogus"])
+            result = runner.invoke(app, ["config", "validate"])
+            assert result.exit_code == 1
+            assert "bad token" in result.output
+
+
+@patch("clickup.cli.commands.config.get_provider")
+def test_config_validate_exception(mock_client_cls):
+    mock_client_cls.side_effect = Exception("network down")
+
+    with tempfile.TemporaryDirectory() as tmpdir:
+        with patch("clickup.core.config.Path.home", return_value=Path(tmpdir)):
+            runner.invoke(app, ["config", "set-token", "pk_test"])
+            result = runner.invoke(app, ["config", "validate"])
+            assert result.exit_code == 1
+            assert "network down" in result.output
+
+
+@patch("clickup.cli.commands.config.get_provider")
+def test_config_whoami_authenticated(mock_client_cls):
+    mock_client = AsyncMock()
+    mock_client.get_user.return_value = User(id=1, username="evan", email="e@x.com", color="#fff")
+    mock_client_cls.return_value = make_mock_ctx(mock_client)
+
+    with tempfile.TemporaryDirectory() as tmpdir:
+        with patch("clickup.core.config.Path.home", return_value=Path(tmpdir)):
+            runner.invoke(app, ["config", "set-token", "pk_test"])
+            result = runner.invoke(app, ["config", "whoami"])
+            assert result.exit_code == 0
+            assert "evan" in result.output
+
+
+def test_config_whoami_no_credentials():
+    with tempfile.TemporaryDirectory() as tmpdir:
+        with patch("clickup.core.config.Path.home", return_value=Path(tmpdir)):
+            with patch.dict("os.environ", {}, clear=False):
+                from os import environ
+
+                for k in [k for k in environ if k.startswith("CLICKUP_")]:
+                    environ.pop(k, None)
+                result = runner.invoke(app, ["config", "whoami"])
+                assert result.exit_code == 1
+                assert "credentials" in result.output.lower()
+
+
+# =============================================================================
+# no-credentials paths — from test_more_coverage.py
+# =============================================================================
+
+
+def test_workspace_list_no_credentials():
+    """Without a token, workspace list refuses with a setup hint."""
+    with tempfile.TemporaryDirectory() as tmpdir:
+        with patch("clickup.core.config.Path.home", return_value=Path(tmpdir)):
+            from os import environ
+
+            for k in [k for k in environ if k.startswith("CLICKUP_")]:
+                environ.pop(k, None)
+            result = runner.invoke(app, ["workspace", "list"])
+            assert result.exit_code != 0
+            assert "API token" in result.output
+
+
+def test_task_list_no_credentials():
+    with tempfile.TemporaryDirectory() as tmpdir:
+        with patch("clickup.core.config.Path.home", return_value=Path(tmpdir)):
+            from os import environ
+
+            for k in [k for k in environ if k.startswith("CLICKUP_")]:
+                environ.pop(k, None)
+            result = runner.invoke(app, ["task", "list", "--list-id", "L1"])
+            assert result.exit_code != 0
